+++ src/core/common/public-api.ts
@@ -22,8 +22,11 @@
 
 export * from './apply-styles-directive';
 export * from './auto-focus.directive';
+export * from './build-string-identifier';
 export * from './common-module';
+export * from './context';
 export * from './dnd-directive';
 export * from './format-if-number';
+export * from './string-identifier';
 export * from './translate-if-string';
 export * from './video.directive';