+++ tools/public_api_guard/core/common.d.ts
@@ -3,6 +3,10 @@ export declare class AjfCommonModule {
     static ɵmod: i0.ɵɵNgModuleDefWithMeta<AjfCommonModule, [typeof i1.AjfDndDirective, typeof i2.AjfVideoDirective, typeof i3.ApplyStylesDirective, typeof i4.AutofocusDirective, typeof i5.FormatIfNumber, typeof i6.TranslateIfString], never, [typeof i1.AjfDndDirective, typeof i2.AjfVideoDirective, typeof i3.ApplyStylesDirective, typeof i4.AutofocusDirective, typeof i5.FormatIfNumber, typeof i6.TranslateIfString]>;
 }
 
+export declare type AjfContext = {
+    [key: string]: any;
+};
+
 export declare class AjfDndDirective {
     file: Observable<FileList>;
     get over(): boolean;
@@ -11,6 +15,11 @@ export declare class AjfDndDirective {
     onDrop(evt: DragEvent): void;
     static ɵdir: i0.ɵɵDirectiveDefWithMeta<AjfDndDirective, "[ajfDnd]", never, {}, { "file": "file"; }, never>;
     static ɵfac: i0.ɵɵFactoryDef<AjfDndDirective, never>;
+}
+
+export interface AjfStringIdentifier {
+    label: string;
+    value: string[];
 }
 
 export declare class AjfVideoDirective implements AfterViewInit {
@@ -38,6 +47,17 @@ export declare class AutofocusDirective implements AfterContentInit {
     ngAfterContentInit(): void;
     static ɵdir: i0.ɵɵDirectiveDefWithMeta<AutofocusDirective, "[autoFocus]", never, { "appAutoFocus": "appAutoFocus"; }, {}, never>;
     static ɵfac: i0.ɵɵFactoryDef<AutofocusDirective, never>;
+}
+
+export declare const buildStringIdentifier: (stringIdentifier: AjfStringIdentifier[] | undefined, context: AjfContext, opts?: BuildStringIdentifierOpts | undefined) => string;
+
+export declare const buildStringIdentifierOpts: (opts?: BuildStringIdentifierOpts | undefined) => Required<BuildStringIdentifierOpts>;
+
+export interface BuildStringIdentifierOpts {
+    emptyString?: string;
+    entriesDivider?: string;
+    labelSuffix?: string;
+    valuesDivider?: string;
 }
 
 export declare class FormatIfNumber extends DecimalPipe {