+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## 11.0.5 "gentle-band" (2020-12-04)
+
+### core
+
+|            |                       |
+| ---------- | --------------------- |
+| feature |  **forms:** add pipe to build string from string identifier ([3fa1ceb5717cb39b1db3cfeb19ab4155e6c3d4be](https://github.com/gnucoop/ajf/commit/3fa1ceb5717cb39b1db3cfeb19ab4155e6c3d4be)) |
+| feature |  **reports:** add pipe to build string from string identifier ([2c585f5a71340f4b50ec1aa24264f0452973ddfa](https://github.com/gnucoop/ajf/commit/2c585f5a71340f4b50ec1aa24264f0452973ddfa)) |
+
+
+## 11.0.4 "bitter-dream" (2020-12-04)
+
+### core
+
+|            |                       |
+| ---------- | --------------------- |
+| bug fix |  **common:** buildStringIdentifier is not exported ([a34307bfa0bbd47a86320fa5c6247c8aa1c7a32a](https://github.com/gnucoop/ajf/commit/a34307bfa0bbd47a86320fa5c6247c8aa1c7a32a)) |
+
+
+## 11.0.3 "late-sound" (2020-12-03)
+
+### core
+
+|            |                       |
+| ---------- | --------------------- |
+| bug fix |  **reports:** widget wrapper inside export container has zero height ([e34d94044084cbfad02f9a018bcd7c2f68ae5266](https://github.com/gnucoop/ajf/commit/e34d94044084cbfad02f9a018bcd7c2f68ae5266)) |
+| feature |  **common:** add helper to build a string from a string identifier ([b2f03b11038fd9f8cd4904334925eb9bc57fc793](https://github.com/gnucoop/ajf/commit/b2f03b11038fd9f8cd4904334925eb9bc57fc793)) |
+| feature |  **reports:** add string identifier in report interface ([6e3abab375de3aaaa86c082bb5acefaf233e98e0](https://github.com/gnucoop/ajf/commit/6e3abab375de3aaaa86c082bb5acefaf233e98e0)) |
+
+
 ## 11.0.2 "shy-term" (2020-11-30)
 
 ### ajf